+++ dbms/src/Functions/FunctionsRegexpCommon.cpp
@@ -1,0 +1,34 @@
+// Copyright 2023 PingCAP, Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+
+#include <Functions/FunctionsRegexpCommon.h>
+
+namespace DB
+{
+template class ParamInt<true>;
+template class ParamInt<false>;
+
+template class ParamString<true>;
+template class ParamString<false>;
+
+template class Param<ParamString<true>, false>;
+template class Param<ParamString<false>, true>;
+template class Param<ParamString<true>, true>;
+template class Param<ParamString<false>, false>;
+template class Param<ParamInt<true>, false>;
+template class Param<ParamInt<false>, true>;
+template class Param<ParamInt<true>, true>;
+template class Param<ParamInt<false>, false>;
+} // namespace DB
