+++ dbms/src/Functions/FunctionsRegexpUndef.h
@@ -1,0 +1,41 @@
+// Copyright 2023 PingCAP, Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+namespace DB
+{
+#undef GET_ACTUAL_INT_PARAM
+#undef GET_ACTUAL_STRING_PARAM
+#undef REGEXP_CLASS_MEM_FUNC_IMPL_NAME
+#undef REPL_PARAM_PTR_VAR_NAME
+#undef MATCH_TYPE_PARAM_PTR_VAR_NAME
+#undef RET_OP_PARAM_PTR_VAR_NAME
+#undef OCCUR_PARAM_PTR_VAR_NAME
+#undef POS_PARAM_PTR_VAR_NAME
+#undef PAT_PARAM_PTR_VAR_NAME
+#undef EXPR_PARAM_PTR_VAR_NAME
+
+#undef REPL_PV_VAR_NAME
+#undef MATCH_TYPE_PV_VAR_NAME
+#undef RET_OP_PV_VAR_NAME
+#undef OCCUR_PV_VAR_NAME
+#undef POS_PV_VAR_NAME
+#undef PAT_PV_VAR_NAME
+#undef EXPR_PV_VAR_NAME
+
+#undef ACTUAL_PARAM_TYPE
+#undef GET_ACTUAL_PARAM_PTR
+#undef ENUMERATE_PARAM_VARIANT_CASES
+#undef APPLY_FOR_PARAM_INT_VARIANTS
+#undef APPLY_FOR_PARAM_STRING_VARIANTS
+} // namespace DB
